feat: v4 finalizado - foi feito migrations - v5 prox

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('file_id')->unsigned();
+            $table->bigInteger('file_id')->unsigned();
             $table->integer('profile')->default(0);
             $table->string('name');
             $table->string('gender')->nullable();
